webapi/tct-audio-html5-tests/audio/audio_play_webm_vorbis.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this lis
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work withou
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
         Zhang, Zhiqiang <zhiqiang.zhang@intel.com>
         Ge, Qing <qingx.ge@intel.com>

-->

<html>
  <head>
    <title>Audio Test: audio_play_webm_vorbis</title>
    <link rel="author" title="Intel" href="http://www.intel.com/"/>
    <link rel="help"href="http://www.w3.org/TR/2012/WD-html5-20121025/media-elements.html#htmlmediaelement"/>
    <meta name="flags" content="" />
    <meta name="assert" content="Check if user agent is able to play audio stream webm vorbis with event ended fired when ended playback"/>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <p><a href="http://www.w3.org/TR/html5/the-iframe-element.html#mediaevents">spec reference</a></p>
    <h4 id="link2">webm/vorbis</h4>
    <audio id="a" controls>
    </audio>
    <div id="log"></div>
    <script>
        var time = 30000;
        var time2 = 1000000;
        var timeout2 = 30000;
        var t = async_test("audio.play()/pause() - to verify if user agent is able to play audio stream webm vorbis with event ended fired when ended playback", {timeout: time});
        setup({timeout:time2});

        var a = document.getElementById("a");

        a.addEventListener('ended', t.step_func(function () {
            t.done();
        }), false);

        a.addEventListener('loadedmetadata', function () {
            a.currentTime = a.duration - 2;
        }, false);

        setTimeout(function () {
            t.step(function () {
                assert_true(false, "there is no trigger event");
            });
            t.done();
        }, timeout2);

        a.src="media/movie_5.webm";

        a.play();
    </script>
  </body>
</html>
